[Moved brush scripts to editor folder]
